+++ services/liaradb/recovery/span/fragment.go
@@ -32,9 +32,10 @@ func newFragment(header []byte, data []byte) *Fragment {
 	}
 }
 
-func (f Fragment) length() int64 { return f.buffer.Length() }
-func (f Fragment) Count() int16  { return f.count.Get() }
-func (f Fragment) Index() int16  { return f.index.Get() }
+// TODO: Fix this cast
+func (f Fragment) length() int  { return int(f.buffer.Length()) }
+func (f Fragment) Count() int16 { return f.count.Get() }
+func (f Fragment) Index() int16 { return f.index.Get() }
 
 func (f Fragment) valid() bool {
 	return page.RestoreCRC(f.crc.Get()).
